feat(vmi-details): refactor vmi details page
