+++ assets/translations/en-us.yaml
@@ -479,6 +479,7 @@ vm:
       console: Console
       networkInterfaces: Network Interfaces
       disks: Disks
+      basics: Basics
     details:
       title:
         vmDetails: Virtual Machine Details
