+++ detail/kubevirt.io.virtualmachine/tabs/details/events.vue
