+++ detail/kubevirt.io.virtualmachine/tabs/details/index.vue
@@ -1,34 +1,26 @@
 <script>
-import { POD } from '@/config/types';
-import CreateEditView from '@/mixins/create-edit-view';
+import HStack from '@/components/Layout/Stack/HStack';
+import VStack from '@/components/Layout/Stack/VStack';
 import VmState from '@/components/formatter/vmState';
-import IPAddress from '@/components/formatter/ipAddress';
-import ResourceState from '../../resource-state/index';
-import LabelsModal from '../../labels-modal';
-import AnnotationsModal from '../../annotations-modal';
-import DescriptionModal from '../../description-modal';
-
-const UNDEFINED = 'n/a';
+import OverviewBasics from './basics';
+import OverviewConfigurations from './configurations';
+import OverviewEvents from './events';
+import OverviewInventory from './inventory';
 
 export default {
-  name: 'Details',
+  name: 'Overview',
 
   components: {
-    ResourceState,
+    HStack,
+    VStack,
     VmState,
-    IPAddress,
-    LabelsModal,
-    AnnotationsModal,
-    DescriptionModal,
+    OverviewBasics,
+    OverviewConfigurations,
+    OverviewEvents,
+    OverviewInventory,
   },
 
-  mixins: [CreateEditView],
-
   props: {
-    value: {
-      type:     Object,
-      required: true
-    },
     resource: {
       type:     Object,
       required: true,
@@ -36,113 +28,33 @@ export default {
         return {};
       }
     },
+    events: {
+      type:     Array,
+      required: true
+    },
     mode: {
       type:     String,
       required: true,
     },
-  },
-
-  fetch() {
-    this.getPods();
+    value: {
+      type:     Object,
+      required: true,
+    }
   },
 
   data() {
-    return {
-      allPods: [],
-
-      labelsModalShow:      false,
-      annotationsModalShow: false,
-      descriptionModalShow: false,
-    };
+    return {};
   },
 
   computed: {
-    name() {
-      return this.value?.metadata?.name || UNDEFINED;
-    },
-    namespace() {
-      return this.value?.metadata?.namespace || UNDEFINED;
-    },
-    creationTimestamp() {
-      const date = new Date(this.value?.metadata?.creationTimestamp);
-
-      return `${ date.getMonth() + 1 }/${ date.getDate() }/${ date.getUTCFullYear() }`;
-    },
-    node() {
-      return this.resource?.status?.nodeName || UNDEFINED;
-    },
-    hostname() {
-      return this.resource?.spec?.hostname || UNDEFINED;
-    },
-    pod() {
-      const pod = this.allPods.find((p) => {
-        return p?.labels['vm.kubevirt.io/name'] === this.name;
-      });
-
-      return pod?.metadata?.name || UNDEFINED;
-    },
-    labelsCount() {
-      const count = Object.keys(this.value?.metadata?.labels).length || 0;
-      const unit = count > 1 ? 'Labels' : 'Label';
-
-      return `${ count } ${ unit }`;
-    },
-    annotationsCount() {
-      const count = Object.keys(this.value?.metadata?.annotations).length || 0;
-      const unit = count > 1 ? 'Annotations' : 'Annotation';
-
-      return `${ count } ${ unit }`;
-    },
-    disks() {
-      const disks = this.value?.spec?.template?.spec?.domain?.devices?.disks || [];
-
-      return disks.filter((disk) => {
-        return !!disk.bootOrder;
-      }).sort((a, b) => {
-        if (a.bootOrder < b.bootOrder) {
-          return -1;
-        }
-
-        return 1;
-      });
-    },
-    flavor() {
-      const domain = this.value?.spec?.template?.spec?.domain;
-
-      return `${ domain.cpu.cores } vCPU , ${ domain.resources.requests.memory } Memory`;
-    },
-    isNamespace() {
-      return 'Namespace';
-    },
-    isPod() {
-      return 'Pod';
-    },
-    isNode() {
-      return 'Node';
-    },
     isDown() {
       return this.isEmpty(this.resource);
     },
   },
 
   methods: {
-    async getPods() {
-      const pods = await this.$store.dispatch('cluster/findAll', { type: POD });
-
-      this.allPods = pods || [];
-    },
-    toggleLabelsModal(show) {
-      this.labelsModalShow = show;
-    },
-    toggleAnnotationsModal(show) {
-      this.annotationsModalShow = show;
-    },
-    toggleDescriptionModal(show) {
-      this.descriptionModalShow = show;
-    },
-    done() {},
-    update() {
-      this.save(() => {});
+    showKeyModal() {
+      this.$refs['overviewInventory'].toggleModal(true);
     },
     isEmpty(o) {
       return o !== undefined && Object.keys(o).length === 0;
@@ -152,328 +64,86 @@ export default {
 </script>
 
 <template>
-  <div class="overview-details">
-    <h2>{{ t("vm.detail.details.title.vmDetails") }}</h2>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.name") }}
-          </label>
-          <div>
-            <span>{{ name }}</span>
-          </div>
+  <HStack class="vm-overview">
+    <VStack class="vm-overview__left">
+      <el-card class="box-card">
+        <div slot="header" class="clearfix">
+          <span>Basics</span>
         </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.status") }}
-          </label>
-          <div>
-            <VmState v-model="value.id" :row="value" />
-          </div>
+        <div>
+          <OverviewBasics v-model="value" :resource="resource" mode="view" />
         </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.namespace") }}
-          </label>
-          <div>
-            <ResourceState v-model="isNamespace" />{{ namespace }}
-          </div>
+      </el-card>
+    </VStack>
+    <VStack class="vm-overview__center">
+      <el-card class="box-card">
+        <div slot="header" class="clearfix">
+          <span>Status</span>
         </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.pod") }}
-          </label>
-          <div>
-            <ResourceState v-model="isPod" />{{ pod }}
-          </div>
+        <div>
+          <VmState v-model="value.id" :row="value" />
         </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.labels") }}
-          </label>
-          <div>
-            <a href="javascript:void(0)" @click="toggleLabelsModal(true)">
-              {{ labelsCount }} <span class="icon icon-edit"></span>
-            </a>
-            <LabelsModal
-              :spec="value"
-              :mode="mode"
-              :visible="labelsModalShow"
-              @close="toggleLabelsModal(false)"
-              @update="update"
-            />
-          </div>
+      </el-card>
+      <el-card class="box-card">
+        <div slot="header" class="clearfix">
+          <span>Configurations</span>
         </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.bootOrder") }}
-          </label>
-          <div>
-            <ul>
-              <li v-for="(disk) in disks" :key="disk.bootOrder">
-                {{ disk.bootOrder }}. {{ disk.name }} (Disk)
-              </li>
-            </ul>
-          </div>
+        <div>
+          <OverviewConfigurations v-model="value" :resource="resource" :mode="mode" />
         </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.annotations") }}
-          </label>
-          <div>
-            <a href="javascript:void(0)" @click="toggleAnnotationsModal(true)">
-              {{ annotationsCount }} <span class="icon icon-edit"></span>
-            </a>
-            <AnnotationsModal
-              :spec="value"
-              :mode="mode"
-              :visible="annotationsModalShow"
-              @close="toggleAnnotationsModal(false)"
-              @update="update"
-            />
-          </div>
+      </el-card>
+      <el-card class="box-card">
+        <div slot="header" class="clearfix">
+          <el-button style="float: right; padding: 3px 0" type="text" icon="el-icon-view" @click="showKeyModal()">
+            SSH-keys
+          </el-button>
+          <span>Inventory</span>
         </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.CDROMs") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
+        <div>
+          <OverviewInventory ref="overviewInventory" v-model="value" />
         </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.description") }}
-          </label>
-          <div>
-            <a href="javascript:void(0)" @click="toggleDescriptionModal(true)">
-              {{ value.metadata.annotations.description }} <span class="icon icon-edit"></span>
-            </a>
-            <DescriptionModal
-              :spec="value"
-              :mode="mode"
-              :visible="descriptionModalShow"
-              @close="toggleDescriptionModal(false)"
-              @update="update"
-            />
-          </div>
+      </el-card>
+    </VStack>
+    <VStack v-if="!isDown" class="vm-overview__right">
+      <el-card class="box-card events-card">
+        <div slot="header" class="clearfix">
+          <span>Events</span>
         </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.ipAddress") }}
-          </label>
-          <div>
-            <IPAddress v-model="value.id" :row="value" />
-          </div>
+        <div class="">
+          <OverviewEvents :events="events" />
         </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.operatingSystem") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.hostname") }}
-          </label>
-          <div v-if="!isDown">
-            {{ hostname }}
-          </div>
-          <div v-else>
-            {{ t("vm.detail.details.down") }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.template") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.timeZone") }}
-          </label>
-          <div v-if="!isDown">
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-          <div v-else>
-            {{ t("vm.detail.details.down") }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.created") }}
-          </label>
-          <div>
-            {{ creationTimestamp }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.node") }}
-          </label>
-          <div>
-            <ResourceState v-model="isNode" />{{ node }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.owner") }}
-          </label>
-          <div>
-            {{ t("vm.detail.noOwner") }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.workloadProfile") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <el-divider></el-divider>
-    <h2>{{ t("vm.detail.details.title.requirements") }}</h2>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.nodeSelector") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.flavor") }}
-          </label>
-          <div>
-            {{ flavor }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.tolerations") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.dedicatedResources") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="row">
-      <div class="col span-6">
-        <div class="labeled-input view">
-          <label>
-            {{ t("vm.detail.details.affinityRules") }}
-          </label>
-          <div>
-            {{ t("vm.detail.notAvailable") }}
-          </div>
-        </div>
-      </div>
-    </div>
-    <el-divider></el-divider>
-    <h2>{{ t("vm.detail.details.title.services") }}</h2>
-    <el-divider></el-divider>
-    <h2>{{ t("vm.detail.details.title.users") }}</h2>
-  </div>
+      </el-card>
+    </VStack>
+  </HStack>
 </template>
 
-<style lang="scss" scoped>
-  .overview-details {
-    display: grid;
-    grid-template-columns: 100%;
-    grid-template-rows: auto;
-    grid-row-gap: 15px;
-
-    h2 {
-      margin-bottom: 15px;
+<style lang="scss">
+  .vm-overview {
+    &__left, &__right {
+      width: 25%;
     }
 
-    ul {
-      padding-left: 0px;
-      list-style-type: none;
+    &__center {
+      flex: 1;
     }
 
-    .badge-state {
-      padding: 2px 5px;
-      font-size: 12px;
-      margin-right: 3px;
+    &__left, &__center, &__right {
+        margin-right: 15px;
+      .el-card {
+        margin-bottom: 15px;
+      }
+      &:last-child, .el-card:last-child {
+        margin: 0;
+      }
+    }
+
+    .el-card__header {
+      font-size: 16px;
+    }
+
+    .events-card .el-card__body {
+      max-height: 500px;
+      overflow: auto;
     }
   }
 </style>
