--- detail/kubevirt.io.virtualmachine/tabs/disks/index.vue
@@ -1,17 +1,0 @@
-<script>
-export default {
-  name:       'Disks',
-  components: {},
-  props:      {},
-  data() {
-    return {};
-  },
-  computed: {},
-};
-</script>
-
-<template>
-  <div></div>
-</template>
-
-<style lang="scss" scoped></style>
